add spark_list_catalog_table_names

+++ create_spark_session.py
@@ -5,6 +5,10 @@ from typing import Any
 
 from pyspark.sql.session import SparkSession
 
+from spark_pipeline_framework.utilities.spark_data_frame_helpers import (
+    spark_list_catalog_table_names,
+)
+
 # make sure env variables are set correctly
 if "SPARK_HOME" not in os.environ:
     os.environ["SPARK_HOME"] = "/usr/local/opt/spark"
@@ -22,12 +26,14 @@ def clean_spark_dir() -> None:
     :return:
     """
     try:
-        os.remove("./derby.log")
-        shutil.rmtree("./metastore_db")
-        shutil.rmtree("./spark-warehouse")
+        if os.path.exists("./derby.log"):
+            os.remove("./derby.log")
+        if os.path.exists("./metastore_db"):
+            shutil.rmtree("./metastore_db")
+        if os.path.exists("./spark-warehouse"):
+            shutil.rmtree("./spark-warehouse")
     except OSError as e:
         print(f"Error cleaning spark directories: {e.strerror}")
-        pass
 
 
 def clean_spark_session(session: SparkSession) -> None:
@@ -36,11 +42,10 @@ def clean_spark_session(session: SparkSession) -> None:
     :param session:
     :return:
     """
-    tables = session.catalog.listTables("default")
+    table_names = spark_list_catalog_table_names(session)
 
-    for table in tables:
-        table_name = table.name
-        print(f"clear_tables() is dropping table/view: {table.name}")
+    for table_name in table_names:
+        print(f"clear_tables() is dropping table/view: {table_name}")
         # Drop the table if it exists
         if session.catalog.tableExists(f"default.{table_name}"):
             # noinspection SqlNoDataSourceInspection
